ec2/requirements.txt
camelot-py[cv]
boto3
pytest